test: add new tests
